requirements.txt
#server 
psutil==5.9.4
colorlog==6.7.0

#micro framwork
uvicorn==0.20.0
fastapi[all]==0.88.0
fastapi-crudrouter==0.8.5
fastapi-signals==0.1.1
starlette==0.22.0
graphene==3.2.1
aerich==0.7.1

#general
PyYAML==6.0

#management
celery==5.2.7

#sse
sse_starlette==1.2.1
redis==4.2.2
aioredis==2.0.1
sseclient==0.0.27

#test
requests==2.28.2
pytest==7.2.1
httpx==0.23.3